feature: eject controllers
